Create .nvmrc
It looks like we're running tests off Node 16 in Travis, and it supports .nvmrc but not .node-version? https://docs.travis-ci.com/user/languages/javascript-with-nodejs/#specifying-nodejs-versions-using-nvmrc

I don't want us to maintain two files with our node versions and I saw this issue which described a way to link the two: https://github.com/shadowspawn/node-version-usage/issues/4

I tried that in this commit—let's see if Travis CI likes it!